+++ packages/components/atoms/f-image-tile/src/components/_tests/ImageTile.test.js
@@ -169,15 +169,16 @@ describe('ImageTile', () => {
             });
         });
 
-        describe('performance tracking', () => {
-            it('calls the performance tracker with `carouselImageTile` after rendering if prop value exists', async () => {
+        describe('performanceTracker', () => {
+            it('calls the performance tracker with `performanceTrackerLabel` after rendering if prop value exists', async () => {
                 // Arrange
                 const performanceTrackerMock = {
                     time: jest.fn()
                 };
 
                 const propsData = {
-                    performanceTracker: performanceTrackerMock
+                    performanceTracker: performanceTrackerMock,
+                    performanceTrackerLabel: 'carouselImageTile'
                 };
 
                 // Act
@@ -187,7 +188,7 @@ describe('ImageTile', () => {
 
                 // Assert
                 expect(performanceTrackerMock.time).toHaveBeenCalledTimes(1);
-                expect(performanceTrackerMock.time).toHaveBeenCalledWith('carouselImageTile');
+                expect(performanceTrackerMock.time).toHaveBeenCalledWith(propsData.performanceTrackerLabel);
             });
         });
     });